apps/vanna-ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vanna>=0.7.0
groq==0.4.0
psycopg[binary]
python-dotenv==1.0.0
sqlalchemy==2.0.23
